feat: the board on screen, and a move that starts work
board.get, the column and card commands, and card.move — the one place
a step is ever started. Landing on a column with a step opens a run;
landing on one without is just a move. There is no schedule, no retry
and no loop: a failed run leaves the card where it is with the reason
on it.

An irreversible step is confirmed before it runs, so a deploy is never
fired by dropping a card on a lane. Deleting a column that holds cards
is refused with the count, which is what lets the screen ask where they
should go instead of only saying no.

The board draws whatever columns come back, in the order they come
back. A test strips comments and asserts no seeded column name appears
in the screen's code — the first version failed on the prose explaining
the rule, and a guard that fails on its own documentation gets deleted
rather than obeyed.

The TypeScript contract is now generated by a test as well as by a dev
run. Generating only from main meant the frontend types were refreshed
just by someone opening the window, so a command could reach the
backend and never reach the screen.

+++ crates/rpc/src/board.rs
@@ -70,8 +70,11 @@ pub struct Run {
     pub output: Option<String>,
     pub exit_code: Option<i32>,
     pub cost_usd: Option<f64>,
-    pub duration_ms: Option<i64>,
-    pub started_at: i64,
+    /// `f64` and not `i64` throughout, because this crosses into a JavaScript
+    /// number and that is what a JavaScript number is — see
+    /// `Commit::committed_at` for the same reason stated once.
+    pub duration_ms: Option<f64>,
+    pub started_at: f64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, Type)]
